% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialization.R
\name{abundance}
\alias{ab}
\alias{abundance}
\title{Fractional abundance}
\usage{
abundance(..., major = get_iso_opts("default_major"), compound = "",
  notation = get_iso_opts("default_ab_notation"), weight = numeric(),
  single_as_df = FALSE)

ab(..., major = get_iso_opts("default_major"), compound = "",
  notation = get_iso_opts("default_ab_notation"), weight = numeric(),
  single_as_df = FALSE)
}
\arguments{
\item{...}{- numeric vectors (can be named) to turn into isotope abundance objects}

\item{major}{- name of the major isotope in the isotope system [optional],
only of importance if converting from abundance to ratio or delta value,
and want automatic name propagation}

\item{compound}{- name of the compound the isotopic values belong to [optional]}

\item{notation}{- what notation the abundance is in ('raw' values or 'percent'),
see \code{\link{switch_notation}} for details}
}
\description{
Generate an isotope abundance object. See \link{isotopia} for general information on initializing
and converting isotope data objects.
}
\details{
The \code{ab} function is a shorthand for \code{abundance} but otherwise identical.
}
\seealso{
Other isotope data types: \code{\link{delta}};
  \code{\link{ff}}, \code{\link{fractionation_factor}};
  \code{\link{intensity}}; \code{\link{ratio}}
}

