% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.alpha}
\alias{as.alpha}
\alias{frac_factor}
\title{Fractionation factor}
\usage{
as.alpha(iso1, iso2)

frac_factor(iso1, iso2)
}
\arguments{
\item{iso1}{the top compound in the fractionation factor}

\item{iso2}{the bottom compound in the fractionation factor}
}
\value{
isotope \code{\link{alpha}} object if parameters can be converted to a \code{\link{alpha}}, an error otherwise
}
\description{
Calculate/convert to isotope fractionation factors (\code{alpha} values).
}
\details{
The \code{frac_factor(...)} function calculates the fractionation factor between two isotope data objects
(for example two delta values, two epsilons, two ratios, or two alpha values).

\code{as.alpha} is mostly synonymous with \code{frac_fractor} but can additionally be used to perform
conversions such as \code{\link{epsilon}} to \code{\link{alpha}}.

All calculatinos are only permissible if the isotope values have matching
attributes.
}
\note{
Some of the conversions are also implemented in arithmetic shorthand, for example to generate
an alpha value from two ratios, \code{frac_factor(ratio(), ratio())} and
\code{as.alpha(ratio(), ratio())} are the same as \code{ratio() / ratio()}.
See \link{arithmetic} for details.
}
\seealso{
Other data type conversions: \code{\link{as.abundance}};
  \code{\link{as.delta}}; \code{\link{as.epsilon}};
  \code{\link{as.ratio}}
}

