% Generated by roxygen2 (4.0.0): do not edit by hand
\name{frac_factor}
\alias{frac_factor}
\title{Fractionation factor}
\usage{
frac_factor(iso, iso2)
}
\arguments{
\item{iso}{the top compound in the fractionation factor}

\item{iso2}{the bottom compound in the fractionation factor}
}
\value{
an object of type \code{\link{alpha}} value
}
\description{
This function calculates the fractionation factor between two isotope data objects
(for example two delta values, two epsilons, two ratios, or two alpha values)
}
\note{
This calculations is also implemented with an \code{\link{arithmetic}} shorthand.
All calculatinos are only permissible if the isotope values have matching
attributes.
}

