% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fractionation_factor}
\alias{ff}
\alias{fractionation_factor}
\title{Fractionation factor
FIXME, change description
Generate a fractionation factor (alpha value = ratio of two isotope ratios) object. See \link{isotopia} for general information on initializing
and converting isotope data objects.}
\usage{
fractionation_factor(..., major = get_iso_opts("default_major"), ctop = "",
  cbot = "", single_as_df = FALSE)

ff(..., major = get_iso_opts("default_major"), ctop = "", cbot = "",
  single_as_df = FALSE)
}
\arguments{
\item{...}{- numeric vectors (can be named) to turn into alpha values}

\item{major}{- name of the major isotope in the isotope system [optional]}

\item{ctop}{- name of the compound representing the top isotope ratio [optional]}

\item{cbot}{- name of the compound representing the bottom isotope ratio [optional]}
}
\description{
The \code{ff} function is a shorthand for \code{fractionation_factor} but otherwise identical.
}
\seealso{
Other isotope data types: \code{\link{ab}},
  \code{\link{abundance}}; \code{\link{delta}};
  \code{\link{epsilon}}, \code{\link{use_permil}};
  \code{\link{intensity}}; \code{\link{ratio}}
}

