% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribs.R
\docType{methods}
\name{get_value}
\alias{get_value}
\alias{get_value,ANY-method}
\alias{ANY-method}
\alias{get_value,numeric-method}
\alias{numeric-method}
\alias{get_value,Isoval-method}
\alias{Isoval-method}
\alias{get_value,Isosys-method}
\alias{Isosys-method}
\title{Retrieve isotope object's primitive values}
\usage{
get_value(iso, notation = iso@notation)

\S4method{get_value}{ANY}(iso, notation = iso@notation)

\S4method{get_value}{numeric}(iso, notation = NULL)

\S4method{get_value}{Isoval}(iso, notation = iso@notation)

\S4method{get_value}{Isosys}(iso, notation = NULL)
}
\arguments{
\item{notation}{specifiy which notation to return the value in (default is the notation
that the object is in)}
}
\value{
In the case of a single isotope object (Isoval), returns the numeric
vector of raw values stored in the object (same as \code{\link[base]{as.numeric}}). 
In the case of an isotope system (Isosys),
returns the data frame underlying the object with all its isotope value
objects also replaced with their numeric raw values. To just get the data
frame but keep the isotope values intact, use \code{\link{as.data.frame}} instead.
}
\description{
This function returns an isotope object's (single value or isotope system)  
primitive data value(s).
}
\seealso{
\code{\link[base]{as.numeric}}, \code{\link{as.data.frame}}, \code{\link[base]{as.data.frame}} (base method)

Other data type attributes: \code{\link{get_name}},
  \code{\link{get_weighted_value}},
  \code{\link{get_weight}}, \code{\link{set_attrib}},
  \code{\link{switch_notation}}, \code{\link{weight}}
}
\concept{data type attributes}
