% Generated by roxygen2 (4.0.0): do not edit by hand
\name{intensity}
\alias{intensity}
\title{Ion intensity}
\usage{
intensity(..., major = get_iso_opts("default_major"), compound = "",
  unit = "", single_as_df = FALSE)
}
\arguments{
\item{...}{- numeric vectors (can be named) to turn into ion intensity objects}

\item{major}{- name of the major isotope in the isotope system [optional],}

\item{compound}{- name of the compound the isotopic values belong to [optional]}

\item{unit}{- units of the measurement (e.g. #, V, mV)}
}
\description{
Generate an ion intensity object (e.g. ion counts or signal intensity).
See \link{isotopia} for general information on initializing and converting isotope data objects.
}
\seealso{
Other isotope data types: \code{\link{ab}},
  \code{\link{abundance}}; \code{\link{delta}};
  \code{\link{ff}}, \code{\link{fractionation_factor}};
  \code{\link{ratio}}
}

