% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_iso_opts}
\alias{set_iso_opts}
\alias{get_iso_opts}
\title{Isotopia options}
\usage{
set_iso_opts(default_ab_notation = c("raw", "percent"),
  default_ff_notation = c("alpha", "eps", "permil", "ppm"),
  default_delta_notation = c("raw", "permil", "ppm"),
  default_intensity_unit = "", default_major = "", default_minor = "",
  exact_mass_balance = FALSE, standards = c())

get_iso_opts(opts)
}
\arguments{
\item{default_ab_notation}{default notation of abundance objects,
see \code{\link{switch_notation}} for details}

\item{default_ff_notation}{default notation of fractionation factors}

\item{default_delta_notation}{default notation of delta values}

\item{default_intensity_unit}{default unit for intensity values}

\item{default_major}{default major isotope on all isotope objects}

\item{exact_mass_balance}{NOT IMPLEMENTED YET!
If enabled, mass balance calculations with delta values (i.e. \code{mass_balance(delta, delta, delta...} 
or \code{\link{delta}() + \link{delta}()}) will always
be performed exact by converting to natural abundances first and making the addition in 
abundance space (will be converted back to delta value afterwards). 
This is only possible if the \code{ref_ratio} in the delta values is
set and will lead to an error if attempted without the reference ratios set.

If disabled, mass balance calculations with delta values (\code{\link{delta}() + \link{delta}()}) will
be performed in delta space (which is not exact but the discrepancy is negligible unless
the minor isotopes in an isotope system make up a significant portion)

see \code{\link{register_standard}} for details}

\item{standards}{isotope ratio objects to register as standards}

\item{default}{minor default minor isotope on all isotope objects}
}
\description{
This allows specifying and retrieving default values for newly
created isotopia objects.
}
\details{
\code{get_iso_opts} allows retrieval of all or individual isotopia options.
Returns a single value if only one option is requested, a named list if multiple
}
\note{
Default options are the following and are set during package loading together
with the default standards
\code{\cr
set_iso_opts(\cr
   default_ab_notation = "raw", \cr
   default_ff_notation = "alpha", \cr
   default_delta_notation = "permil", \cr
   default_intensity_unit = "", \cr
   default_major = "", \cr
   default_minor = "",\cr
   exact_mass_balance = FALSE\cr
)
}
}
\examples{
get_iso_opts("standards") # get a table of all standards
get_iso_opts(c("default_major", "default_minor")) # get a named list with the 
default major and minor isotopes
}
\seealso{
Other options: \code{\link{register_standard}}
}
