% Generated by roxygen2 (4.0.0): do not edit by hand
\name{label}
\alias{label}
\title{Get the full label of an isotope data object}
\usage{
label(object)
}
\description{
Get the full label of an isotope data object
}
\examples{
\dontrun{
label(ratio(...))
label(abundance(...))
label(isosys(ratio(...), ratio(...))
}
}

