% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\docType{methods}
\name{mass_balance}
\alias{mass_balance}
\alias{mass_balance,ANY,ANY-method}
\alias{ANY-method}
\alias{mass_balance,Abundance,Abundance-method}
\alias{Abundance,Abundance-method}
\alias{mass_balance,Delta,Delta-method}
\alias{Delta,Delta-method}
\alias{mass_balance,Deltas,Deltas-method}
\alias{Deltas,Deltas-method}
\title{Calculate isotope mass balance}
\usage{
mass_balance(iso, iso2, ..., exact = get_iso_opts("exact_mass_balance"))

\S4method{mass_balance}{ANY,ANY}(iso, iso2, ...,
  exact = get_iso_opts("exact_mass_balance"))

\S4method{mass_balance}{Abundance,Abundance}(iso, iso2, ...,
  exact = get_iso_opts("exact_mass_balance"))

\S4method{mass_balance}{Delta,Delta}(iso, iso2, ...,
  exact = get_iso_opts("exact_mass_balance"))

\S4method{mass_balance}{Deltas,Deltas}(iso, iso2, ...,
  exact = get_iso_opts("exact_mass_balance"))
}
\arguments{
\item{...}{- any number of weighted isotope value objects (have to be all either abundance or delta)}

\item{exact}{- whether to calculate mass balance of delta values exactly (default FALSE), not
fully implemented yet}
}
\value{
weighted abundance or delta value object that represents the combination of the parameters
}
\description{
This function calculates the isotope mass balance from combining multiple weighted
isotope \code{\link{abundance}} or \code{\link{delta}} value objects. This calculation
is also implemented with an \code{\link{arithmetic}} shorthand.
}
\seealso{
Other operations: \code{\link{fractionate}},
  \code{\link{shift_reference}}
}
