% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{get_name}
\alias{get_name}
\alias{get_name,ANY-method}
\alias{ANY-method}
\alias{get_name,Isoval-method}
\alias{Isoval-method}
\alias{get_name,Ratio-method}
\alias{Ratio-method}
\alias{get_name,Abundance-method}
\alias{Abundance-method}
\alias{get_name,FractionationFactor-method}
\alias{FractionationFactor-method}
\alias{get_name,Delta-method}
\alias{Delta-method}
\alias{get_units}
\alias{get_units,Isoval-method}
\alias{get_units,Intensity-method}
\alias{Intensity-method}
\alias{get_label}
\alias{get_label,ANY-method}
\alias{get_label,Isoval-method}
\alias{get_label,FractionationFactor-method}
\alias{get_label,Delta-method}
\alias{get_label,Isosys-method}
\alias{Isosys-method}
\title{Information about an isotopic data object}
\usage{
get_name(object)

\S4method{get_name}{ANY}(object)

\S4method{get_name}{Isoval}(object)

\S4method{get_name}{Ratio}(object)

\S4method{get_name}{Abundance}(object)

\S4method{get_name}{FractionationFactor}(object)

\S4method{get_name}{Delta}(object)

get_units(object)

\S4method{get_units}{Isoval}(object)

\S4method{get_units}{Intensity}(object)

get_label(object)

\S4method{get_label}{ANY}(object)

\S4method{get_label}{Isoval}(object)

\S4method{get_label}{FractionationFactor}(object)

\S4method{get_label}{Delta}(object)

\S4method{get_label}{Isosys}(object)
}
\description{
Get information about the name, label and units of
an isotopic data object.
}
\details{
\code{get_name()} returns the name of an isotopic data object

\code{get_units()} provides the units of an isotope data object depending
on the object type and notation

\code{get_label()} provides the full label of an isotope data object
}
\examples{
\dontrun{
get_label(ratio(...))
get_label(abundance(...))
get_label(isosys(ratio(...), ratio(...))
}
}
\seealso{
Other data type attributes: \code{\link{get_value}},
  \code{\link{get_weighted_value}},
  \code{\link{get_weight}}, \code{\link{set_attrib}},
  \code{\link{switch_notation}}, \code{\link{weight}}
}
\concept{data type attributes}
