% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{quietly}
\alias{quietly}
\title{Run a calculation quietly.}
\usage{
quietly(expr)
}
\description{
This small utility function is just a convenient wrapper for running
isotope calculations silently without outputting any of the warnings or
messages (it uses \link[base]{suppressMessages} and \link[base]{suppressWarnings}
internally) that might occur. Use with care to suppress warnings, you
might end up hiding important information.
}

