% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,Isoval-method}
\alias{show,Isoval-method}
\alias{Isoval-method}
\alias{show,Isosys-method}
\alias{Isosys-method}
\title{show information}
\usage{
\S4method{show}{Isoval}(object)

\S4method{show}{Isosys}(object)
}
