% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\docType{methods}
\name{switch_notation}
\alias{switch_notation}
\alias{switch_notation,ANY,ANY,ANY-method}
\alias{ANY-method}
\alias{switch_notation,numeric,Notation_raw,Notation_raw-method}
\alias{numeric,Notation_raw,Notation_raw-method}
\alias{switch_notation,numeric,Notation_percent,Notation_percent-method}
\alias{numeric,Notation_percent,Notation_percent-method}
\alias{switch_notation,numeric,Notation_alpha,Notation_alpha-method}
\alias{numeric,Notation_alpha,Notation_alpha-method}
\alias{switch_notation,numeric,Notation_eps,Notation_eps-method}
\alias{numeric,Notation_eps,Notation_eps-method}
\alias{switch_notation,numeric,Notation_permil,Notation_permil-method}
\alias{numeric,Notation_permil,Notation_permil-method}
\alias{switch_notation,numeric,Notation_ppm,Notation_ppm-method}
\alias{numeric,Notation_ppm,Notation_ppm-method}
\alias{switch_notation,numeric,Notation_percent,Notation_raw-method}
\alias{numeric,Notation_percent,Notation_raw-method}
\alias{switch_notation,numeric,Notation_raw,Notation_percent-method}
\alias{numeric,Notation_raw,Notation_percent-method}
\alias{switch_notation,numeric,Notation_eps,Notation_raw-method}
\alias{numeric,Notation_eps,Notation_raw-method}
\alias{switch_notation,numeric,Notation_raw,Notation_eps-method}
\alias{numeric,Notation_raw,Notation_eps-method}
\alias{switch_notation,numeric,Notation_eps,Notation_alpha-method}
\alias{numeric,Notation_eps,Notation_alpha-method}
\alias{switch_notation,numeric,Notation_alpha,Notation_eps-method}
\alias{numeric,Notation_alpha,Notation_eps-method}
\alias{switch_notation,numeric,Notation_alpha,Notation_raw-method}
\alias{numeric,Notation_alpha,Notation_raw-method}
\alias{switch_notation,numeric,Notation_permil,Notation_raw-method}
\alias{numeric,Notation_permil,Notation_raw-method}
\alias{switch_notation,numeric,Notation_raw,Notation_permil-method}
\alias{numeric,Notation_raw,Notation_permil-method}
\alias{switch_notation,numeric,Notation_permil,Notation_eps-method}
\alias{numeric,Notation_permil,Notation_eps-method}
\alias{switch_notation,numeric,Notation_eps,Notation_permil-method}
\alias{numeric,Notation_eps,Notation_permil-method}
\alias{switch_notation,numeric,Notation_permil,Notation_alpha-method}
\alias{numeric,Notation_permil,Notation_alpha-method}
\alias{switch_notation,numeric,Notation_alpha,Notation_permil-method}
\alias{numeric,Notation_alpha,Notation_permil-method}
\alias{switch_notation,numeric,Notation_ppm,ANY-method}
\alias{numeric,Notation_ppm,ANY-method}
\alias{switch_notation,numeric,ANY,Notation_ppm-method}
\alias{numeric,Notation_ANY,Notation_ppm-method}
\alias{switch_notation,Isoval,ANY,missing-method}
\alias{Isoval,ANY,missing-method}
\alias{switch_notation,Isoval,Notation_raw,missing-method}
\alias{Isoval,Notation_raw,missing-method}
\alias{switch_notation,Abundance,Notation_percent,missing-method}
\alias{Abundance,Notation_percent,missing-method}
\alias{switch_notation,Delta,Notation_permil,missing-method}
\alias{Delta,Notation_permil,missing-method}
\alias{switch_notation,Delta,Notation_ppm,missing-method}
\alias{Delta,Notatation_ppm,missing-method}
\alias{switch_notation,FractionationFactor,Notation_alpha,missing-method}
\alias{FractionationFactor,Notation_alpha,missing-method}
\alias{switch_notation,FractionationFactor,Notation_eps,missing-method}
\alias{FractionationFactor,Notation_eps,missing-method}
\alias{switch_notation,FractionationFactor,Notation_raw,missing-method}
\alias{FractionationFactor,Notation_raw,missing-method}
\alias{switch_notation,FractionationFactor,Notation_permil,missing-method}
\alias{FractionationFactor,Notation_permil,missing-method}
\alias{switch_notation,FractionationFactor,Notation_ppm,missing-method}
\alias{FractionationFactor,Notation_ppm,missing-method}
\alias{switch_notation,Isoval,character,missing-method}
\alias{Isoval,character,missing-method}
\alias{switch_notation,Isosys,character,missing-method}
\alias{Isosys,character,missing-method}
\title{Switch notation}
\usage{
switch_notation(iso, to)

\S4method{switch_notation}{ANY,ANY,ANY}(iso, to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_raw,Notation_raw}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{numeric,Notation_percent,Notation_percent}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_alpha,Notation_alpha}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_eps,Notation_eps}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{numeric,Notation_permil,Notation_permil}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_ppm,Notation_ppm}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{numeric,Notation_percent,Notation_raw}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_raw,Notation_percent}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_eps,Notation_raw}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{numeric,Notation_raw,Notation_eps}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{numeric,Notation_eps,Notation_alpha}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_alpha,Notation_eps}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_alpha,Notation_raw}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_permil,Notation_raw}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_raw,Notation_permil}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_permil,Notation_eps}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_eps,Notation_permil}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_permil,Notation_alpha}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_alpha,Notation_permil}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{numeric,Notation_ppm,ANY}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{numeric,ANY,Notation_ppm}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{Isoval,ANY,missing}(iso, to = NULL, from = NULL)

\S4method{switch_notation}{Isoval,Notation_raw,missing}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{Abundance,Notation_percent,missing}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{Delta,Notation_permil,missing}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{Delta,Notation_ppm,missing}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{FractionationFactor,Notation_alpha,missing}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{FractionationFactor,Notation_eps,missing}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{FractionationFactor,Notation_raw,missing}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{FractionationFactor,Notation_permil,missing}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{FractionationFactor,Notation_ppm,missing}(iso,
  to = NULL, from = NULL)

\S4method{switch_notation}{Isoval,character,missing}(iso, to = NULL,
  from = NULL)

\S4method{switch_notation}{Isosys,character,missing}(iso, to = NULL,
  from = NULL)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ff}}, \code{\link{abundance}}, \code{\link{delta}})}

\item{to}{which notation to convert to}
}
\value{
isotope object with converted notation, an error if it is not a valid conversion
}
\description{
Convert from one notation to another for an isotope data object.
}
\details{
Valid notations depend on the data type: 
\itemize{
 \item{\code{\link{abundance}}: }{'raw', 'percent'}
 \item{\code{\link{delta}}: }{'raw', 'permil', 'ppm'}
 \item{\code{\link{fractionation_factor}}: }{'alpha', 'eps', 'permil', 'ppm'}
}
}
\seealso{
Other data type attributes: \code{\link{get_name}},
  \code{\link{get_value}},
  \code{\link{get_weighted_value}},
  \code{\link{get_weight}}, \code{\link{set_attrib}},
  \code{\link{weight}}
}
