% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{to_abundance}
\alias{to_abundance}
\alias{to_ab}
\alias{to_abundance,ANY-method}
\alias{ANY-method}
\alias{to_abundance,Abundance-method}
\alias{Abundance-method}
\alias{to_abundance,Abundances-method}
\alias{Abundances-method}
\alias{to_abundance,Ratio-method}
\alias{Ratio-method}
\alias{to_abundance,Ratios-method}
\alias{Ratios-method}
\alias{to_abundance,Intensities-method}
\alias{Intensities-method}
\alias{to_abundance,Delta-method}
\alias{Delta-method}
\title{Convert to isotope abundance}
\usage{
to_abundance(iso)

to_ab(iso)

\S4method{to_abundance}{ANY}(iso)

\S4method{to_abundance}{Abundance}(iso)

\S4method{to_abundance}{Abundances}(iso)

\S4method{to_abundance}{Ratio}(iso)

\S4method{to_abundance}{Ratios}(iso)

\S4method{to_abundance}{Intensities}(iso)

\S4method{to_abundance}{Delta}(iso)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.)}
}
\value{
isotope \code{\link{abundance}} object if iso can be converted to a \code{\link{abundance}}, an error otherwise
}
\description{
\code{to_abundance} converts another isotopic data type to an abundance.
The \code{to_ab} function is a shorthand for \code{to_abundance} but otherwise identical.
}
\seealso{
Other data type conversions: \code{\link{to_delta}},
  \code{\link{to_ff}}, \code{\link{to_ratio}}
}
\concept{data type conversions}
