% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{to_delta}
\alias{to_delta}
\alias{to_d}
\alias{to_delta,ANY,ANY-method}
\alias{ANY-method}
\alias{to_delta,Isosys,missing-method}
\alias{Isosys,missing-method}
\alias{to_delta,Delta,missing-method}
\alias{Delta,missing-method}
\alias{to_delta,FractionationFactor,missing-method}
\alias{FractionationFactor,missing-method}
\alias{to_delta,FractionationFactor,numeric-method}
\alias{FractionationFactor,numeric-method}
\alias{to_delta,FractionationFactor,Ratio-method}
\alias{FractionationFactor,Ratio-method}
\alias{to_delta,Ratio,missing-method}
\alias{Ratio,missing-method}
\alias{to_delta,Ratio,numeric-method}
\alias{Ratio,numeric-method}
\alias{to_delta,Ratio,Ratio-method}
\alias{Ratio,Ratio-method}
\alias{to_delta,Ratios,Ratios-method}
\alias{Ratios,Ratios-method}
\alias{to_delta,Abundance,ANY-method}
\alias{Abundance,ANY-method}
\title{Convert to delta value}
\usage{
to_delta(iso, ref_ratio)

to_d(iso, ref_ratio)

\S4method{to_delta}{ANY,ANY}(iso, ref_ratio)

\S4method{to_delta}{Isosys,missing}(iso, ref_ratio)

\S4method{to_delta}{Delta,missing}(iso, ref_ratio)

\S4method{to_delta}{FractionationFactor,missing}(iso, ref_ratio)

\S4method{to_delta}{FractionationFactor,numeric}(iso, ref_ratio)

\S4method{to_delta}{FractionationFactor,Ratio}(iso, ref_ratio)

\S4method{to_delta}{Ratio,missing}(iso, ref_ratio)

\S4method{to_delta}{Ratio,numeric}(iso, ref_ratio)

\S4method{to_delta}{Ratio,Ratio}(iso, ref_ratio)

\S4method{to_delta}{Ratios,Ratios}(iso, ref_ratio)

\S4method{to_delta}{Abundance,ANY}(iso, ref_ratio)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.)}

\item{ref_ratio}{the refernce ratio associated with the delta value. This is optional but required if planning
later conversions back to ratios or abundane values. Can be supplied as a raw numeric numer or a Ratio object
(in the case of the latter, the compound name of the Ratio object will be registered as the name of the
reference).}
}
\value{
isotope \code{\link{delta}} object if iso can be converted to a \code{\link{delta}}, an error otherwise
}
\description{
\code{to_delta} converts another isotopic data type to a delta value.
The \code{to_d} function is a shorthand for \code{to_delta} but otherwise identical.
}
\seealso{
Other data type conversions: \code{\link{to_abundance}},
  \code{\link{to_ff}}, \code{\link{to_ratio}}
}
