% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{to_ff}
\alias{to_ff}
\alias{to_ff,ANY,ANY-method}
\alias{ANY-method}
\alias{to_ff,FractionationFactor,missing-method}
\alias{FractionationFactor,missing-method}
\alias{to_ff,FractionationFactors,missing-method}
\alias{FractionationFactors,missing-method}
\alias{to_ff,Ratio,Ratio-method}
\alias{Ratio,Ratio-method}
\alias{to_ff,Delta,missing-method}
\alias{Delta,missing-method}
\alias{to_ff,Delta,Delta-method}
\alias{Delta,Delta-method}
\title{Fractionation factor}
\usage{
to_ff(iso1, iso2)

\S4method{to_ff}{ANY,ANY}(iso1, iso2)

\S4method{to_ff}{FractionationFactor,missing}(iso1, iso2)

\S4method{to_ff}{FractionationFactors,missing}(iso1, iso2)

\S4method{to_ff}{Ratio,Ratio}(iso1, iso2)

\S4method{to_ff}{Delta,missing}(iso1, iso2)

\S4method{to_ff}{Delta,Delta}(iso1, iso2)
}
\arguments{
\item{iso1}{the top compound in the fractionation factor}

\item{iso2}{the bottom compound in the fractionation factor}
}
\value{
isotope \code{\link{fraction_factor}} object if parameters can be converted, an error otherwise
}
\description{
Calculate/convert to an isotope \code{\link{fractionation_factor}}
}
\details{
The \code{to_ff(...)} function calculates the fractionation factor between two isotope data objects
(for example two \code{\link{delta}} values, two \code{\link{ratio}}, or two \code{\link{ff}}).
All calculatinos are only permissible if the isotope values have matching
attributes and fractionation factors will be returend in the default notation
(see \code{\link{set_iso_opts}} for details)
}
\note{
Some of the conversions are also implemented in arithmetic shorthand, for example to generate
an fractionation factor in alpha notation from two ratios 
\code{to_ff(ratio(), ratio())} is the same as \code{ratio() / ratio()}.
See \link{arithmetic} for details.
}
\seealso{
Other data type conversions: \code{\link{to_abundance}},
  \code{\link{to_delta}}, \code{\link{to_ratio}}
}
