% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to_ratio}
\alias{to_ratio}
\title{Convert to isotope ratio}
\usage{
to_ratio(iso)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.)}
}
\value{
isotope \code{\link{ratio}} object if iso can be converted to a \code{\link{ratio}}, an error otherwise
}
\description{
\code{to_ratio} converts another isotopic data type to a ratio.
}
\seealso{
Other data type conversions: \code{\link{frac_factor}},
  \code{\link{to_ff}}; \code{\link{to_abundance}};
  \code{\link{to_delta}}; \code{\link{to_epsilon}}
}

