% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attribs.R
\name{update_iso}
\alias{update_iso}
\title{update_iso the attributes of an isotope value object
internal function that is called by set_attrib wrapper}
\usage{
update_iso(obj, attribs)
}
\description{
update_iso the attributes of an isotope value object
internal function that is called by set_attrib wrapper
}

