% Generated by roxygen2 (4.0.0): do not edit by hand
\name{arithmetic}
\alias{arithmetic}
\title{Isotope arithmetic}
\usage{
intensity +- intensity

abundance +- abundance

delta +- delta

ff - 1

ff * ratio

delta * delta

intensity / intensity

ratio / ratio

alpha / alpha

delta / delta
}
\description{
Several arithmetic operators (+, -, *, /) are implemented to work with specific
isotope value object to allow shorthand data type conversions and calculations.
Operations are generally only permitted if the two isotope objects being combined
have matching attributes (isotope name, major isotope, etc.).
}
\details{
\code{intensity+-intensity} allows the addition of intensity values, the result is a another \code{\link{intensity}} object

\code{abundance+-abundance} is a shorthand for calculating the isotopic mass balance of
two \code{\link{abundance}} objects, see \code{\link{mass_balance}} for details

\code{delta+-delta} is a shorthand for calculating the isotopic mass balance of
two \code{\link{delta}} objects, see \code{\link{mass_balance}} for details,
will use the global \code{\link{exact_mass_balance}} setting for the calculation

\code{ff - 1} is a shorthand for converting a fractionation factor from
alpha to epsilon notation. The ff object has to be in alpha notation,
otherwise this is just interpreted as a regular arithmetic operation
and the result will no longer be an isotope object.

\code{ff*ratio}, \code{ff*ff}, \code{ff*delta} are a shorthand for
fractionating an isotope object with a factionation factor,
see \code{\link{fractionate}} for details

\code{delta*delta}, is a shorthand for shifting the reference frame of the
first delta value to that of the second (requires the compound measured in the
second to be the reference of the first!), see \code{\link{shift_reference}}
for details

\code{intensity/intensity} allows the creation of an isotope \code{\link{ratio}} object

\code{ratio/ratio} allows the creation of an isotope \code{\link{alpha}} object (a fractionation factor).
This is a shorthand for the \link{frac_factor} function.

\code{alpha/ alpha} allows the creation of another isotope \code{\link{alpha}} object but requires that
either the denominator names or numerator names of the two alpha objects are identical (i.e. they "cancel").
 This is a shorthand for the \link{frac_factor} function.

\code{delta/delta} creates an \code{\link{alpha}} fractionation factor object that
describes the fractionation factor between the two compounds, requires the reference
name of the two delta values to be identical. This is a shorthand for the \link{frac_factor} function.
Works identically for two epsilon values.
}

