% Generated by roxygen2 (4.0.0): do not edit by hand
\name{delta}
\alias{delta}
\title{Delta value}
\usage{
delta(..., major = get_iso_opts("default_major"), compound = "", ref = "",
  ref_ratio = numeric(), notation = get_iso_opts("default_delta_notation"),
  weight = numeric(), single_as_df = FALSE)
}
\arguments{
\item{...}{- numeric vectors (can be named) to turn into delta values}

\item{major}{- name of the major isotope in the isotope system [optional]}

\item{compound}{- name of the compound the isotopic values belong to [optional]}

\item{ref}{- name of the reference material}

\item{ref_ratio}{- value of the reference material}

\item{notation}{- which notation the value is in, \code{"permil"} (1000x multiplicaiton),
\code{"raw"} (raw value, i.e. no multiplication) and \code{"ppm"} (10^6 multiplicaiton)
are currently implemented for \code{delta} values. See \code{\link{switch_notation}} on
details how to convert between notations.}

\item{weight}{- weight the isotope value (with a mass, concentration, etc.) for easy mass balance calculations.
The default value is 1, i.e. an unweighted isotope value.
If specified, \code{weight} must be a single value or a numeric vector of the same size as the data values.
The weight of an isotope value obejct can be retrieved and (re)set with the \code{\link{weight}} function.}
}
\description{
Generate an isotope delta value object. See \link{isotopia} for general information on initializing
and converting isotope data objects. Delta values can be easily converted from values in one notation to values in
another notation by using \code{\link{switch_notation}}.
}
\details{
For mass balance calculations with delta values, simply add the appropriate weights (if different from
the default) and use \code{delta(...) + delta(...)}.
}
\examples{
delta(50, notation = "permil") # enter as permil value
delta(0.05, notation = "raw") # enter as non-permil value
}
\seealso{
Other isotope data types: \code{\link{ab}},
  \code{\link{abundance}}; \code{\link{ff}},
  \code{\link{fractionation_factor}};
  \code{\link{intensity}}; \code{\link{ratio}}
}

