% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\docType{methods}
\name{fractionate}
\alias{fractionate}
\alias{fractionate,ANY,ANY-method}
\alias{ANY-method}
\alias{fractionate,FractionationFactor,Ratio-method}
\alias{FractionFactor,Ratio-method}
\alias{fractionate,FractionationFactor,FractionationFactor-method}
\alias{FractionFactor,FractionationFactor-method}
\alias{fractionate,FractionationFactor,Delta-method}
\alias{FractionFactor,Delta-method}
\alias{fractionate,Delta,FractionationFactor-method}
\alias{Delta,FractionationFactor-method}
\title{Fractionate an isotopic value}
\usage{
fractionate(frac, delta)
fractionate(delta, frac)
fractionate(frac, ratio)

\S4method{fractionate}{ANY,ANY}(frac, iso)

\S4method{fractionate}{FractionationFactor,Ratio}(frac, iso)

\S4method{fractionate}{FractionationFactor,FractionationFactor}(frac, iso)

\S4method{fractionate}{FractionationFactor,Delta}(frac, iso)

\S4method{fractionate}{Delta,FractionationFactor}(frac, iso)
}
\arguments{
\item{frac}{the fractionation factor \code{\link{ff}} used to fractionate the isotope value}

\item{delta}{isotope object to fractionate}

\item{ratio}{isotope object to fractionate}
}
\value{
an object of the same type as the isotope object that gets fractionated
}
\description{
This function calculates the outcome of isotopic fractionation by a \code{\link{fractionation_factor}}
and can be applied to \code{\link{ratio}} data, \code{\link{delta}} values or other
\code{\link{fractionation_factor}} objects.
}
\note{
Several of these calculations are also
implemented with an \code{\link{arithmetic}} shorthand. All calculatinos are
only permissible if the fractionation factors and isotope values have matching
attributes.
}
\seealso{
Other operations: \code{\link{mass_balance}},
  \code{\link{shift_reference}}
}
\concept{operations}
