% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialization.R
\name{fractionation_factor}
\alias{ff}
\alias{fractionation_factor}
\title{Fractionation factor}
\usage{
fractionation_factor(..., major = get_iso_opts("default_major"),
  notation = get_iso_opts("default_ff_notation"), ctop = "", cbot = "",
  single_as_df = FALSE)

ff(..., major = get_iso_opts("default_major"),
  notation = get_iso_opts("default_ff_notation"), ctop = "", cbot = "",
  single_as_df = FALSE)
}
\arguments{
\item{...}{- numeric vectors (can be named) to turn into fractionation factors}

\item{major}{- name of the major isotope in the isotope system [optional]}

\item{notation}{- which notation the value is in, \code{"alpha"} (alpha value), \code{"eps"} (epsilon value),
\code{"permil"} (epsilon * 1000) are currently implemented for \code{fractionation_factor} values.
See \code{\link{switch_notation}} on details how to convert between notations.}

\item{ctop}{- name of the compound representing the top isotope ratio [optional]}

\item{cbot}{- name of the compound representing the bottom isotope ratio [optional]}
}
\description{
Generate a fractionation factor object. Can be initialized in \code{alpha} and
\code{epsilon} notation. Fractionation factors can be easily converted from values in one notation to values in
another notation by using \code{\link{switch_notation}}.
}
\details{
See \link{isotopia} for general information on initializing
and converting to other isotope data objects.
The \code{ff} function is a shorthand for \code{fractionation_factor} but otherwise identical.
}
\seealso{
Other isotope data types: \code{\link{ab}},
  \code{\link{abundance}}; \code{\link{delta}};
  \code{\link{intensity}}; \code{\link{ratio}}
}

