% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mass_balance}
\alias{mass_balance}
\title{Calculate isotope mass balance}
\usage{
mass_balance(iso, iso2, ..., exact = exact_mass_balance())
}
\arguments{
\item{...}{- any number of weighted isotope value objects (have to be all either abundance or delta)}

\item{exact}{- whether to calculate mass balance of delta values exactly (default FALSE),
use exact_mass_balance to set this paramter globally}
}
\value{
weighted abundance or delta value object that represents the combination of the parameters
}
\description{
This function calculates the isotope mass balance from combining multiple weighted
isotope \code{\link{abundance}} or \code{\link{delta}} value objects. This calculation
is also implemented with an \code{\link{arithmetic}} shorthand.
}

