% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.R
\name{get_name}
\alias{get_label}
\alias{get_name}
\alias{get_units}
\title{Information about an isotopic data object}
\usage{
get_name(object)

get_units(object)

get_label(object)
}
\description{
Get information about the name, label and units of
an isotopic data object.
}
\details{
\code{get_name()} returns the name of an isotopic data object

\code{get_units()} provides the units of an isotope data object depending
on the object type and notation

\code{get_label()} provides the full label of an isotope data object
}
\examples{
\dontrun{
get_label(ratio(...))
get_label(abundance(...))
get_label(isosys(ratio(...), ratio(...))
}
}
\seealso{
Other data type attributes: \code{\link{get_value}};
  \code{\link{get_weighted_value}};
  \code{\link{get_weight}}; \code{\link{set_attrib}};
  \code{\link{switch_notation}}; \code{\link{weight}}
}

