% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{recast_isoval}
\alias{recast_isoval}
\title{generic function to recast an isotopic value object during conversions}
\usage{
recast_isoval(iso, to_class, mods = list(), validate = TRUE)
}
\arguments{
\item{iso}{object}

\item{to_class}{- which class to cast to}

\item{mods}{- list of modifications to existing attributes (can be list(x = NULL) for removing attribute x)}

\item{validate}{- whether to validate after the recast, default TRUE}
}
\description{
generic function to recast an isotopic value object during conversions
}

