% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{register_standard}
\alias{register_standard}
\alias{get_standards}
\alias{get_standard}
\title{Isotope standards}
\usage{
register_standard(ratio)

get_standards(minor = NULL, major = NULL, name = NULL)

get_standard(minor = NULL, major = NULL, name = NULL)
}
\arguments{
\item{ratio}{- a \code{\link{ratio}} object with 'minor', 
and 'major' isotope as well as 'compound' (the name of the standard) attributes all defined}

\item{minor}{- character vector of minor isotope names to search for}

\item{major}{- character vector of major isotope names to search for}

\item{name}{- character vector of standards names to search for}
}
\value{
list of ratio objects
}
\description{
Isotopia provides functionality to register and retrieve isotope standards.
Registered standards can be used for automatic conversions of, for example,
delta values which have attributes that match a standard.
}
\details{
Use \code{register_standard()} to register an isotope standard. This can be useful
for keeping track of standards you use internally and will also allow
conversions from \code{\link{delta}} to e.g. \code{\link{ratio}} to 
automatically try to find the approriate standard for the conversion
from the registered values.

Use \code{get_stanards} to retrieve any number of registered isotope standards
that can be identified with the provided search terms. For an overview
table of all standards (rather than the actual ratio objects), 
use \code{get_iso_opts("standards")} instead.

\code{get_standard} is the same as \code{get_standards} except that it
returns a single object from the found standards and throws an error
if the search criteria did not yield exactly one.
}
\seealso{
Other options: \code{\link{set_iso_opts}}
}
\concept{options}
