% Generated by roxygen2 (4.0.0): do not edit by hand
\name{switch_notation}
\alias{switch_notation}
\title{Switch notation}
\usage{
switch_notation(iso, to)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ff}}, \code{\link{abundance}}, \code{\link{delta}})}
}
\value{
isotope object with converted notation, an error if it is not a valid conversion
}
\description{
Valid notation:
Abundance: raw, percent
Delta: raw, permil, ppm
Fractionation factor: alpha, eps, permil, ppm (note: alpha is considered the 'raw' unit of a fractionation factor)
}
\seealso{
Other data type attributes: \code{\link{get_value}};
  \code{\link{get_weighted_value}};
  \code{\link{get_weight}}; \code{\link{set_attrib}};
  \code{\link{weight}}
}

