% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to_abundance}
\alias{to_abundance}
\title{Convert to isotope abundance}
\usage{
to_abundance(iso)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.)}
}
\value{
isotope \code{\link{abundance}} object if iso can be converted to a \code{\link{abundance}}, an error otherwise
}
\description{
\code{to_abundance} converts another isotopic data type to an abundance.
}
\seealso{
Other data type conversions: \code{\link{frac_factor}},
  \code{\link{to_ff}}; \code{\link{to_delta}};
  \code{\link{to_epsilon}}; \code{\link{to_ratio}}
}

