% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to_abundance}
\alias{to_ab}
\alias{to_abundance}
\title{Convert to isotope abundance}
\usage{
to_abundance(iso)

to_ab(iso)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.)}
}
\value{
isotope \code{\link{abundance}} object if iso can be converted to a \code{\link{abundance}}, an error otherwise
}
\description{
\code{to_abundance} converts another isotopic data type to an abundance.
The \code{to_ab} function is a shorthand for \code{to_abundance} but otherwise identical.
}
\seealso{
Other data type conversions: \code{\link{to_d}},
  \code{\link{to_delta}}; \code{\link{to_ff}};
  \code{\link{to_r}}, \code{\link{to_ratio}}
}

