% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to_delta}
\alias{to_d}
\alias{to_delta}
\title{Convert to delta value}
\usage{
to_delta(iso, ref_ratio, permil = use_permil())

to_d(iso, ref_ratio, permil = use_permil())
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.)}

\item{ref_ratio}{the refernce ratio associated with the delta value. This is optional but required if planning
later conversions back to ratios or abundane values. Can be supplied as a raw numeric numer or a Ratio object
(in the case of the latter, the compound name of the Ratio object will be registered as the name of the
reference).}

\item{permil}{whether to generate epsilon object in permil values (x1000) or not}
}
\value{
isotope \code{\link{delta}} object if iso can be converted to a \code{\link{delta}}, an error otherwise
}
\description{
\code{to_delta} converts another isotopic data type to a delta value.
The \code{to_d} function is a shorthand for \code{to_delta} but otherwise identical.
}
\seealso{
Other data type conversions: \code{\link{frac_factor}},
  \code{\link{to_ff}}; \code{\link{to_ab}},
  \code{\link{to_abundance}}; \code{\link{to_epsilon}};
  \code{\link{to_r}}, \code{\link{to_ratio}}
}

