% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to_epsilon}
\alias{to_epsilon}
\title{Convert to epsilon value}
\usage{
to_epsilon(iso, permil = use_permil())
}
\arguments{
\item{iso}{isotopic data object (e.g. \code{\link{alpha}})}

\item{permil}{whether to generate epsilon object in permil values (x1000) or not}
}
\value{
isotope \code{\link{epsilon}} object if iso can be converted to an \code{\link{epsilon}}, an error otherwise
}
\description{
\code{to_delta} converts another isotopic data type to an epsilon values
}
\seealso{
Other data type conversions: \code{\link{to_ab}},
  \code{\link{to_abundance}}; \code{\link{to_d}},
  \code{\link{to_delta}}; \code{\link{to_ff}};
  \code{\link{to_r}}, \code{\link{to_ratio}}
}

