% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to_ff}
\alias{to_ff}
\title{Fractionation factor}
\usage{
to_ff(iso1, iso2, notation = "alpha")
}
\arguments{
\item{iso1}{the top compound in the fractionation factor}

\item{iso2}{the bottom compound in the fractionation factor}
}
\value{
isotope \code{\link{fraction_factor}} object if parameters can be converted, an error otherwise
}
\description{
Calculate/convert to isotope fractionation factors (\code{ff} values).
}
\details{
FIXME
The \code{frac_factor(...)} function calculates the fractionation factor between two isotope data objects
(for example two delta values, two epsilons, two ratios, or two alpha values).


All calculatinos are only permissible if the isotope values have matching
attributes.
}
\note{
Some of the conversions are also implemented in arithmetic shorthand, for example to generate
an fractionation factor in alpha notation from two ratios
\code{to_ff(ratio(), ratio())} is the same as \code{ratio() / ratio()}.
See \link{arithmetic} for details.
}
\seealso{
Other data type conversions: \code{\link{to_ab}},
  \code{\link{to_abundance}}; \code{\link{to_d}},
  \code{\link{to_delta}}; \code{\link{to_r}},
  \code{\link{to_ratio}}
}

