% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{to_ratio}
\alias{to_ratio}
\alias{to_r}
\alias{to_ratio,ANY-method}
\alias{ANY-method}
\alias{to_ratio,Ratio-method}
\alias{Ratio-method}
\alias{to_ratio,Ratios-method}
\alias{Ratios-method}
\title{Convert to isotope ratio}
\usage{
to_ratio(iso)

to_r(iso)

\S4method{to_ratio}{ANY}(iso)

\S4method{to_ratio}{Ratio}(iso)

\S4method{to_ratio}{Ratios}(iso)
}
\arguments{
\item{iso}{isotopic data object (\code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.)}
}
\value{
isotope \code{\link{ratio}} object if iso can be converted to a \code{\link{ratio}}, an error otherwise
}
\description{
\code{to_ratio} converts another isotopic data type to a ratio. 
The \code{to_r} function is a shorthand for \code{to_ratio} but otherwise identical.
}
\seealso{
Other data type conversions: \code{\link{to_abundance}},
  \code{\link{to_delta}}, \code{\link{to_ff}}
}
\concept{data type conversions}
