% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribs.R
\docType{methods}
\name{weight}
\alias{weight}
\alias{weight,ANY,ANY-method}
\alias{ANY-method}
\alias{weight,Isoval,numeric-method}
\alias{Isovale,numeric-method}
\title{Weight an isotope value object}
\usage{
weight(iso, weight)

\S4method{weight}{ANY,ANY}(iso, weight)

\S4method{weight}{Isoval,numeric}(iso, weight)
}
\arguments{
\item{iso}{- object to get weight or add weight}

\item{weight}{- vector of weight values, has to be a single value or the same length
as the data stored in the isotope value object.}
}
\description{
\code{weight(iso, weight)} adds a weight (can be thought of as mass or concentration) to an isotopic value
which will be used to weigh the isotope value when adding together multiple isotope values. 
\code{\link{get_weight}(iso)} returns the weight of an isotope value object.
}
\note{
This can also be achieved when first initializing (or updating) an object 
via calls to \code{\link{ratio}}, \code{\link{abundance}}, \code{\link{delta}}, etc.
}
\examples{
r <- ratio(0.2)
r <- weight(r, 10)
print(get_weight(r)) # returns 10
}
\seealso{
Other data type attributes: \code{\link{get_name}},
  \code{\link{get_value}},
  \code{\link{get_weighted_value}},
  \code{\link{get_weight}}, \code{\link{set_attrib}},
  \code{\link{switch_notation}}
}
\concept{data type attributes}
